sun: wait for fs_ready entry to be created before writing
Sometimes the sys fs entry '/sys/kernel/cnss/fs_ready'
is not created when trying to write to it on
post-fs-data, then the write will fail and so WLAN CBC
which depends on it won't be triggered.

To fix it, wait for at most 1 seconds before writing to
the entry.

Change-Id: If07397c7fce49df6d38641548b2ceea88c550c3d
CRs-Fixed: 3917119